+++ src/Component/JsonSchema/Printer.php
@@ -97,7 +97,6 @@ EOH
         $command = new FixCommand(new ToolInfo());
         $config = [
             'path' => [$path],
-            '--allow-risky' => true,
         ];
 
         if (!empty($this->fixerConfig)) {