+++ src/Component/JsonSchema/Tests/fixtures/reserved-words/expected/Normalizer/JaneObjectNormalizer.php
@@ -14,7 +14,7 @@ class JaneObjectNormalizer implements DenormalizerInterface, NormalizerInterface
     use DenormalizerAwareTrait;
     use NormalizerAwareTrait;
     use CheckArray;
-    protected $normalizers = array('Jane\\Component\\JsonSchema\\Tests\\Expected\\Model\\_List' => 'Jane\\Component\\JsonSchema\\Tests\\Expected\\Normalizer\\_ListNormalizer', '\\Jane\\Component\\JsonSchemaRuntime\\Reference' => '\\Jane\\Component\\JsonSchema\\Tests\\Expected\\Runtime\\Normalizer\\ReferenceNormalizer'), $normalizersCache = array();
+    protected $normalizers = array('Jane\\Component\\JsonSchema\\Tests\\Expected\\Model\\_List' => 'Jane\\Component\\JsonSchema\\Tests\\Expected\\Normalizer\\ListNormalizer', '\\Jane\\Component\\JsonSchemaRuntime\\Reference' => '\\Jane\\Component\\JsonSchema\\Tests\\Expected\\Runtime\\Normalizer\\ReferenceNormalizer'), $normalizersCache = array();
     public function supportsDenormalization($data, $type, $format = null)
     {
         return array_key_exists($type, $this->normalizers);