+++ src/Component/OpenApi2/Tests/fixtures/exceptions/expected/Client.php
@@ -9,7 +9,7 @@ class Client extends \Jane\Component\OpenApi2\Tests\Expected\Runtime\Client\Clie
      * @throws \Jane\Component\OpenApi2\Tests\Expected\Exception\TestNoTagBadRequestException
      * @throws \Jane\Component\OpenApi2\Tests\Expected\Exception\TestNoTagNotFoundException
      * @throws \Jane\Component\OpenApi2\Tests\Expected\Exception\TestNoTagInternalServerErrorException
-     * @throws \Jane\Component\OpenApi2\Tests\Expected\Exception\TestNoTag600Exception
+     * @throws \Jane\Component\OpenApi2\Tests\Expected\Exception\TestNoTagCustom600Exception
      *
      * @return null|\Psr\Http\Message\ResponseInterface
      */