+++ src/Component/OpenApi3/Tests/fixtures/issue-337/expected/Runtime/Client/Client.php
@@ -47,7 +47,7 @@ abstract class Client
                 $request = $request->withBody($body);
             } elseif (is_resource($body)) {
                 $request = $request->withBody($this->streamFactory->createStreamFromResource($body));
-            } elseif (strlen($body) <= 4000 && is_file($body)) {
+            } elseif (strlen($body) <= 4000 && @file_exists($body)) {
                 // more than 4096 chars will trigger an error
                 $request = $request->withBody($this->streamFactory->createStreamFromFile($body));
             } else {